Move cmd/pivit/status/status.go -> pkg/pivit/status/status.go
