+++ cmd/pivit/verify.go
@@ -8,8 +8,8 @@ import (
 	"io"
 	"os"
 
-	"github.com/cashapp/pivit/cmd/pivit/status"
 	"github.com/cashapp/pivit/cmd/pivit/yubikey"
+	"github.com/cashapp/pivit/pkg/pivit/status"
 	"github.com/cashapp/pivit/pkg/pivit/utils"
 	"github.com/certifi/gocertifi"
 	cms "github.com/github/smimesign/ietf-cms"
